schedule and reschedule screens unit test refactor
